+++ pom.xml
@@ -2,7 +2,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>me.reratos.timehandler</groupId>
 	<artifactId>TimeHandler</artifactId>
-	<version>1.3.0-beta</version>
+	<version>1.3.1-beta</version>
 	
 	<properties>
 		<maven.compiler.target>1.8</maven.compiler.target>
