+++ src/main/java/me/reratos/timehandler/handler/CommandHandler.java
@@ -4,7 +4,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
 
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
@@ -15,10 +14,9 @@ import org.bukkit.entity.Player;
 
 import me.reratos.timehandler.TimeHandler;
 import me.reratos.timehandler.core.TimeManager;
-import me.reratos.timehandler.core.WeatherManager;
 import me.reratos.timehandler.core.WorldManager;
-import me.reratos.timehandler.enums.MoonPhasesEnum;
 import me.reratos.timehandler.handler.commands.HelpCommand;
+import me.reratos.timehandler.handler.commands.InfoCommand;
 import me.reratos.timehandler.handler.commands.SetCommand;
 import me.reratos.timehandler.utils.UpdateChecker;
 
@@ -35,53 +33,7 @@ public class CommandHandler {
 	}
 	
 	public static boolean info(CommandSender sender, World world) {
-		String worldName = world.getName();
-		
-		MemorySection worldsConfigMS = (MemorySection) TimeHandler.worldsConfig.get("worlds." + worldName);
-        
-        if(worldsConfigMS == null) {
-        	TimeHandler.sendMessage(sender, "This world has NOT yet been added to the handler.");
-        } else {
-        	WorldManager wm = TimeManager.getRunablesWorld().get(worldName);
-        	
-        	String status = ChatColor.RESET + "[" + (wm.getWorld() == null ? ChatColor.RED + "RUNNING ERROR" : (
-        			wm.isEnabled() ? ChatColor.GREEN + "RUNNING" : ChatColor.RED + "OFF")) +
-        			ChatColor.RESET + "]";
-        	
-        	sender.sendMessage(ChatColor.YELLOW + "World name: " + ChatColor.GREEN + worldName + 
-        			ChatColor.RESET +" - " + status);
-        	
-        	sender.sendMessage("Current time: " + world.getTime() + ", FullTime: " + world.getFullTime());
-        	
-        	long days = world.getFullTime() / 24000;
-        	int phase = (int) (days % 8);
-        	sender.sendMessage("Moon phase: " + ChatColor.BLUE + MoonPhasesEnum.values()[phase].name());
-        	
-        	LinkedHashMap<String, Object> list = (LinkedHashMap<String, Object>) worldsConfigMS.getValues(true);
-
-        	String climaAtual = WeatherManager.getClimaAtual(world);
-        	sender.sendMessage("Current weather: " + climaAtual + ", change in: " + world.getWeatherDuration());
-        	
-        	// Lista as informações de ambiente do mundo
-        	listInfo(sender, list);
-        }
-        
-		return true;
-	}
-	
-	private static void listInfo(CommandSender sender, Map<String, Object> list) {
-		if(list == null) {
-			return;
-		}
-
-		sendMessage(sender, "enabled: " + getValueMessageInfo(list, "enabled"));
-		sendMessage(sender, "weather: " + getValueMessageInfo(list, "weather"));
-		sendMessage(sender, "thunder: " + getValueMessageInfo(list, "thunder"));
-		sendMessage(sender, "time: " 	+ getValueMessageInfo(list, "time"));
-		sendMessage(sender, "timeFixed: " + getValueMessageInfo(list, "timeFixed"));
-		sendMessage(sender, "durationDay: " + getValueMessageInfo(list, "durationDay"));
-		sendMessage(sender, "durationNight: " + getValueMessageInfo(list, "durationNight"));
-		sendMessage(sender, "moonPhase: " + getValueMessageInfo(list, "moonPhase"));
+		return InfoCommand.commandInfoBase(sender, world);
 	}
 	
 	public static boolean help(CommandSender sender) {
@@ -221,14 +173,4 @@ public class CommandHandler {
 		return list;
 	}
 
-	private static void sendMessage(CommandSender sender, String message) {
-		sender.sendMessage("  " + ChatColor.YELLOW + message);
-	}
-	
-	private static String getValueMessageInfo(Map<String, Object> list, String key) {
-		Object obj = list.get(key);
-		return obj != null ? ChatColor.WHITE + obj.toString() : ChatColor.DARK_GRAY + "-" ;
-	}
-	
-	
 }
