+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: TimeHandler
-version: 1.3.0-beta
+version: 1.3.1-beta
 main: me.reratos.timehandler.TimeHandler
 author: Relry
 database: false
